# Major refactoring going forward
header files moved to includes directory and minor renaming done. Currently the server/client communication, setup and clean-up code is tightly coupled with the battledot client handling, cmd line parsing and game lobby. Manual testing and debugging has worked up until this point but it is time to bring in a testing framework, and refactor to accommodate logging. The DataPacket struct needs to be looked at for serialization, the x and y fields are handled with Network Byte Order but the flags are not and the buffer field may not be necessary at this point.